extract/requirements.txt
psycopg[binary]>=3.1
python-dotenv